requirements.txt
ultralytics
opencv-python
numpy
pandas
streamlit
folium
streamlit-folium
pytz
gpxpy
pyyaml
fastapi
uvicorn
streamlit-webrtc
flask

